% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{call_mirai_}
\alias{call_mirai_}
\title{mirai (Call Value)}
\usage{
call_mirai_(x)
}
\arguments{
\item{x}{a \sQuote{mirai} object.}
}
\description{
\code{call_mirai_} is a variant of \code{call_mirai} that allows user
    interrupts, suitable for interactive use.
}
\note{
Usage is deprecated in favour of \code{x[]}, which also allows user
    interrupts.
}
\keyword{internal}
