% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{cloud_config}
\alias{cloud_config}
\title{Cloud Remote Launch Configuration}
\usage{
cloud_config(platform = "posit")
}
\arguments{
\item{platform}{[default "posit"] character name of the platform
(case-insensitive). Currently the only option is "posit" to use the Posit
Workbench launcher.}
}
\value{
A list in the required format to be supplied to the \code{remote} argument
of \code{\link[=daemons]{daemons()}} or \code{\link[=launch_remote]{launch_remote()}}.
}
\description{
Generates a remote configuration for launching daemons via cloud /
cloud-based managed platforms.
}
\examples{
tryCatch(cloud_config(), error = identity)

\dontrun{

# Launch 2 daemons using the Posit Workbench default:
daemons(n = 2, url = host_url(), remote = cloud_config(platform = "posit"))
}

}
\seealso{
\code{\link[=ssh_config]{ssh_config()}}, \code{\link[=cluster_config]{cluster_config()}}, and \code{\link[=remote_config]{remote_config()}} for other
types of remote launch configuration.
}
