% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{collect_mirai}
\alias{collect_mirai}
\alias{collect_mirai_}
\title{mirai (Collect Value)}
\usage{
collect_mirai(x)

collect_mirai_(x)
}
\arguments{
\item{x}{a \sQuote{mirai} object or list of \sQuote{mirai} objects.}
}
\value{
Depending on the type of \sQuote{x} supplied, either an object or a
    list of objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{collect_mirai} retrieves the value of a single, or list of,
    \sQuote{mirai} objects, waiting for their completion if still in progress.

\code{collect_mirai_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

df1 <- data.frame(a = 1, b = 2)
df2 <- data.frame(a = 3, b = 1)
m <- mirai(as.matrix(rbind(df1, df2)), df1 = df1, df2 = df2, .timeout = 1000)
collect_mirai(m)

ml <- mwalk(c(a = 2, b = 3, c = 4), rnorm, mean = 20, .args = list(sd = 2))
collect_mirai(ml)

}

}
