% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{cpi}
\alias{cpi}
\title{Compute Profile Information}
\usage{
cpi(x = c("n", "pid", "tls", "urls"), .compute = "default")
}
\arguments{
\item{x}{[default 'n'] information to retrieve, one of 'n' (integer number of
daemons set, or the client URL if not using dispatcher), 'pid' (dispatcher
process ID), 'tls' (TLS configuration), or 'urls' (daemon URLs when using
dispatcher).}

\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
List, integer or character vector (as applicable), or else NULL if
    not available.
}
\description{
Retrieve information for the selected compute profile (read-only).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1L)
cpi("n")
cpi("pid")
cpi("urls")
daemons(0)

}

}
