% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{cpinfo}
\alias{cpinfo}
\title{Compute Profile Information}
\usage{
cpinfo(x = c("pid", "tls", "urls"), .compute = "default")
}
\arguments{
\item{x}{[default 'pid'] info to request, one of "pid" (dispatcher process
ID), "tls" (TLS configuration), or "urls" (daemon URLs when using
dispatcher).}

\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
The requested information, or else NULL if not present.
}
\description{
Retrieve information for the selected compute profile (read-only).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1L)
cpinfo("pid")
cpinfo("urls")
daemons(0)

}

}
