% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{daemons_sync}
\alias{daemons_sync}
\title{Synchronous Daemons}
\usage{
daemons_sync(.compute = NULL)
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile
to use (each has its own independent set of daemons), or NULL to use the
'default' profile.}
}
\value{
Invisibly, logical \code{TRUE}.
}
\description{
Set daemons which evaluate mirai synchronously in your current R process.
Specify a value for \code{.compute} to limit synchronous behaviour to the named
compute profile, otherwise all mirai will be performed synchronously by
default.
}
\details{
Using synchronous daemons substantially changes the behaviour of mirai by
evaluating them immediately after mirai creation. This can be useful for
interactive debugging via a \code{\link[=browser]{browser()}} instance in your current session.

Synchronous daemons may be reset in the usual way with a call to \code{daemons(0)}
for the relevant compute profile.
}
\examples{
daemons_sync()
info()
m <- mirai(Sys.getpid())
daemons(0)
m[]

daemons_sync("seq")
info("seq")
with_daemons("seq", {
  m <- mirai(Sys.getpid())
})
daemons(0, .compute = "seq")
m[]

}
\seealso{
\code{\link[=daemons]{daemons()}}.
}
