% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{daemons_view}
\alias{daemons_view}
\title{View Daemons}
\usage{
daemons_view()
}
\value{
A named list comprising: \itemize{
    \item{\code{daemons}} {- integer number of daemons set.}
    \item{\code{connections}} {- integer number of active connections at the
    client URL.}
    }
}
\description{
View the number of currently active 'daemons' or persistent server processes.
}
\details{
Note: for an active queue, the number of connections will always be
    1L as only the queue connects to the client. When using a client URL, the
    number of daemons will always show as 1L, however the connections will
    reflect the number of actual connected servers.
}
\examples{
daemons_view()

}
