% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{dispatcher_v1}
\alias{dispatcher_v1}
\title{Dispatcher (v1)}
\usage{
dispatcher_v1(
  host,
  url = NULL,
  n = NULL,
  ...,
  retry = FALSE,
  token = FALSE,
  tls = NULL,
  pass = NULL,
  rs = NULL,
  monitor = NULL
)
}
\arguments{
\item{host}{the character host URL to dial (where tasks are sent from),
including the port to connect to (and optionally for websockets, a path),
e.g. 'tcp://hostname:5555' or 'ws://10.75.32.70:5555/path'.}

\item{url}{(optional) the character URL dispatcher should listen at,
including the port to connect to (and optionally for websockets, a path),
e.g. 'tcp://hostname:5555' or 'ws://10.75.32.70:5555/path'.
Specify 'tls+tcp://' or 'wss://' to use secure TLS connections. Daemons
should dial in to this URL.}

\item{n}{(optional) if specified, the integer number of daemons to launch. In
this case, a local url is automatically generated.}

\item{...}{(optional) additional arguments passed through to
\code{\link{daemon}}. These include \sQuote{asyncdial}, \sQuote{autoexit},
and \sQuote{cleanup}.}

\item{retry}{[default FALSE] logical value, whether to automatically retry
tasks where the daemon crashes or terminates unexpectedly on the next
daemon instance to connect. If TRUE, the mirai will remain unresolved but
\code{\link{status}} will show \sQuote{online} as 0 and \sQuote{assigned} >
\sQuote{complete}. To cancel a task in this case, use
\code{saisei(force = TRUE)}. If FALSE, such tasks will be returned as
\sQuote{errorValue} 19 (Connection reset).}

\item{token}{[default FALSE] if TRUE, appends a unique 24-character token to
each URL path the dispatcher listens at (not applicable for TCP URLs which
do not accept a path).}

\item{tls}{[default NULL] (required for secure TLS connections)
\strong{either} the character path to a file containing the PEM-encoded TLS
certificate and associated private key (may contain additional certificates
leading to a validation chain, with the TLS certificate first), \strong{or}
a length 2 character vector comprising [i] the TLS certificate (optionally
certificate chain) and [ii] the associated private key.}

\item{pass}{[default NULL] (required only if the private key supplied to
\sQuote{tls} is encrypted with a password) For security, should be provided
through a function that returns this value, rather than directly.}

\item{rs}{[default NULL] the initial value of .Random.seed. This is set
automatically using L'Ecuyer-CMRG RNG streams generated by the host process
and should not be independently supplied.}

\item{monitor}{(for package internal use only) do not set this parameter.}
}
\value{
The regenerated character URL upon success, or else NULL.
}
\description{
This is a deprecated legacy function only present to support existing uses.
}
\keyword{internal}
