% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{.daemon}
\alias{.daemon}
\title{dot Daemon}
\usage{
.daemon(url, exitlinger = 2000L)
}
\arguments{
\item{url}{the character host or dispatcher URL to dial into, including the
port to connect to (and optionally for websockets, a path), e.g.
'tcp://10.75.32.70:5555' or 'ws://10.75.32.70:5555/path'.}

\item{exitlinger}{[default 2000L] time in milliseconds to linger before
exiting to allow the socket to complete sends currently in progress.}
}
\value{
Invisible NULL.
}
\description{
Implements an ephemeral executor for the remote process.
}
\keyword{internal}
