% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{.flat}
\alias{.flat}
\alias{.progress}
\alias{.stop}
\title{mirai Map Options}
\format{
An object of class \code{bytecode} of length 1.

An object of class \code{bytecode} of length 1.

An object of class \code{bytecode} of length 1.
}
\usage{
.flat

.progress

.stop
}
\description{
Expressions to be provided to the \verb{[]} method for 'mirai_map' objects.
}
\section{Collection Options}{


\code{x[]} collects the results of a 'mirai_map' \code{x} and returns a list. This will
wait for all asynchronous operations to complete if still in progress,
blocking but user-interruptible.

\code{x[.flat]} collects and flattens map results to a vector, checking that
they are of the same type to avoid coercion. Note: errors if an 'errorValue'
has been returned or results are of differing type.

\code{x[.progress]} collects map results whilst showing a progress bar from
the \CRANpkg{cli} package, if installed, with completion percentage and ETA,
or else a simple text progress indicator. Note: if the map operation
completes too quickly then the progress bar may not show at all.

\code{x[.stop]} collects map results applying early stopping, which stops at
the first failure and cancels remaining operations. Note: operations already
in progress continue to completion, although their results are not collected.

The options above may be combined in the manner of: \cr
\code{x[.stop, .progress]} which applies early stopping together with a
progress indicator.
}

\keyword{datasets}
\keyword{internal}
