% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{.progress}
\alias{.progress}
\alias{.stop}
\title{mirai Map Options}
\format{
An object of class \code{expression} of length 1.

An object of class \code{expression} of length 1.
}
\usage{
.progress

.stop
}
\description{
Expressions to insert into the \code{[]} method for \sQuote{mirai_map}
    objects.
}
\section{Results}{


    \code{x[]} collects the results of a mirai_map \code{x}. This will wait
    for all asynchronous operations to complete if still in progress,
    blocking but user-interruptible.

    \code{x[.progress]} collects the results whilst showing a text progress
    indicator.

    \code{x[.stop]} collects the results applying early stopping, which stops
    at the first failure and aborts all remaining in-progress operations.

    \code{x[c(.stop, .progress)]} combines early stopping with a progress
    indicator.
}

\keyword{datasets}
\keyword{internal}
