% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{.server}
\alias{.server}
\title{mirai dot Server (Async Executor)}
\usage{
.server(url, exitlinger = 2000L)
}
\arguments{
\item{url}{the client or dispatcher URL to dial into as a character string,
including the port to connect to and (optionally) a path for websocket
URLs e.g. 'tcp://192.168.0.2:5555' or 'ws://192.168.0.2:5555/path'.}

\item{exitlinger}{[default 2000L] time in milliseconds to linger before
exiting to allow the socket to complete sends currently in progress.}
}
\value{
Invisible NULL.
}
\description{
Implements an ephemeral executor/server for the remote process.
}
\keyword{internal}
