% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{info}
\alias{info}
\title{Information}
\usage{
info(.compute = "default", alt = FALSE)
}
\arguments{
\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}

\item{alt}{[default FALSE] returns the connections and daemons status matrix,
\strong{alternatively} the self-signed X.509 certificate and dispatcher
URLs.}
}
\value{
A named list comprising:
    \itemize{
    \item{\strong{connections}} {- number of active connections. Always 1L
    when using dispatcher as there is a single connection to the dispatcher,
    which in turn connects to the daemons.}
    \item{\strong{daemons}} {- if using dispatcher: a status matrix (see
    Status Matrix section below), or else an integer 'errorValue' if
    communication with the dispatcher was unsuccessful. If not using
    dispatcher: the number of daemons set, or else the host URL.}
    }

    Alternatively:
    \itemize{
    \item{\strong{tls}} {- character vector comprising the self-signed X509
    certificate and empty certificate revocation list, if applicable, or else
    NULL.}
    \item{\strong{urls}} {- character vector comprising the URLs at
    dispatcher, if applicable, or else NULL.}
    }
}
\description{
Retrieve information for the specified compute profile, comprising current
    connections and daemons status. Alternatively, the self-signed X.509
    certificate, and URLs at dispatcher.
}
\section{Status Matrix}{


    When using dispatcher, \code{$daemons} comprises a matrix with the
    following columns:
    \itemize{
    \item{\strong{online}} {- shows as 1 when there is an active connection,
    or else 0 if a daemon has yet to connect or has disconnected.}
    \item{\strong{instance}} {- increments by 1 every time there is a new
    connection at a URL. This counter is designed to track new daemon
    instances connecting after previous ones have ended (due to time-outs
    etc.). 'instance' resets to zero if the URL is regenerated by
    \code{\link{saisei}}.}
    \item{\strong{assigned}} {- shows the cumulative number of tasks assigned
    to the daemon.}
    \item{\strong{complete}} {- shows the cumulative number of tasks
    completed by the daemon.}
    }
    The URLs are stored as row names to the matrix.
}

\section{Alternate Information}{


    The return values may be passed directly to the 'url' and 'tls'
    arguments of \code{\link{daemon}} when manually launching remote daemon
    instances. \code{\link{bquote}} may be used to construct the call, as per
    the below example.

    Note: the 'tls' argument does not need to be specified if launching local
    dameons using \code{\link{launch}}, which will supply it automatically.
}

\examples{
if (interactive()) {
# Only run examples in interactive R sessions

info()
daemons(n = 2L, url = "wss://[::1]:0")
info()
info(alt = TRUE)

# creates call to daemon
call <- bquote(daemon(url = .(info(alt = TRUE)$urls[1]), tls = .(info(alt = TRUE)$tls)))
call

daemons(0)

}

}
