% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_error_value}
\alias{is_error_value}
\title{Is Error Value}
\usage{
is_error_value(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'errorValue', FALSE otherwise.
}
\description{
Is the object an error value generated by the system or a 'miraiError' from
    failed execution within a mirai. Includes user-specified errors such as
    mirai timeouts.
}
\examples{
is_error_value(1L)

}
