% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_mirai_interrupt}
\alias{is_mirai_interrupt}
\title{Is mirai Interrupt}
\usage{
is_mirai_interrupt(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'miraiInterrupt', FALSE otherwise.
}
\description{
Is the object a 'miraiInterrupt'. When a mirai is sent a user interrupt, e.g.
    by ctrl+c during an ongoing \code{\link{call_mirai}}, the mirai will
    resolve to an empty character string classed as 'miraiInterrupt' and
    'errorValue'. To test for all errors, including timeouts etc.,
    \code{\link{is_error_value}} should be used instead.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

m <- mirai(stop())
call_mirai(m)
is_mirai_interrupt(m$data)

}

}
