% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{launch}
\alias{launch}
\title{Launch Background Process}
\usage{
launch(args)
}
\arguments{
\item{args}{character string, which will be shell quoted by \code{\link{shQuote}}
before being passed as argument.}
}
\value{
Invisibly, an exit code (zero upon success).
}
\description{
Utility function which calls \code{Rscript} in a background process, with
    'args' passed as command line argument to \code{Rscript -e 'args'}. May
    be used to re-launch local daemons that have timed out, for instance.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

launch('mirai::server("abstract://mirai", idletime = 60000L)')

}

}
