% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{make_cluster}
\alias{make_cluster}
\alias{stop_cluster}
\title{Make Mirai Cluster}
\usage{
make_cluster(
  n,
  url = NULL,
  ssh = list(nodes = character(), port = 22, timeout = 5),
  ...
)

stop_cluster(cl)
}
\arguments{
\item{n}{integer number of nodes (launched on the local machine unless 'url'
is specified).}

\item{url}{(specify for remote nodes), the character URL on the host for
remote nodes to dial into, including a port accepting incoming connections,
e.g. 'tcp://10.75.37.40:5555'. Specify a URL starting 'tls+tcp://' to use
secure TLS connections.}

\item{ssh}{(if 'url' is specified, for launching remote nodes via SSH) a named
list with 'nodes' being a character vector of hostnames or IP addresses
of the remote machines on which to launch nodes, e.g.
\code{c('10.75.37.90', 'nodename')}, and optionally 'port' as the numeric
port number on which to connect [default 22] and 'timeout' as the maximum
time allowed for connection setup in seconds [default 5].}

\item{...}{additional arguments passed onto \code{\link{daemons}}.}

\item{cl}{a 'miraiCluster'.}
}
\value{
For \strong{make_cluster}: An object of class 'miraiCluster' and
    'cluster'. Each 'miraiCluster' has an automatically assigned ID and 'n'
    nodes of class 'miraiNode'.

    For \strong{stop_cluster}: invisible NULL.
}
\description{
\code{make_cluster} creates a cluster of type 'miraiCluster', which may be
    used as a cluster object for any function in the \pkg{parallel} base
    package such as \code{\link[parallel]{clusterApply}} or
    \code{\link[parallel]{parLapply}}.

\code{stop_cluster} stops a cluster created by \code{make_cluster}.
}
\details{
The defaults correspond most closely to existing usage in the
    \pkg{parallel} package, although '...' arguments are passed onto
    \code{\link{daemons}} for additional customisation if desired.

    For remote nodes, the 'ssh' argument is a convenience feature with the
    most common settings. If used, the number of nodes is inferred from the
    length of the character vector 'nodes' and 'n' is disregarded if supplied.

    By specifying 'url' and 'n', nodes may also be launched by alternative
    means, for example using \code{\link{launch_remote}}.

    Note: requires R >= 4.4 (currently R-devel). Methods will not work with
    prior R versions.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

cl <- make_cluster(2)
cl
cl[[1L]]
stop_cluster(cl)

cl <- tryCatch(
  make_cluster(
    # host URL accessible to nodes
    url = "tcp://10.75.37.40:5555",
    # node IP / hostnames to connect to with custom port number
    ssh = list(nodes = c("10.75.37.90", "10.75.37.91"), port = 222)
  )
, error = identity)
if (inherits(cl, "cluster")) stop_cluster(cl)

}

}
