% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mdata}
\alias{mdata}
\alias{mdata_}
\title{mirai (Data)}
\usage{
mdata(x)

mdata_(x)
}
\arguments{
\item{x}{a list of \sQuote{mirai} objects.}
}
\value{
An list, the same length as \sQuote{x}, preserving names.
}
\description{
\code{mdata} retrieves the data of a list of \sQuote{mirai} objects, waiting
    for resolution if still in progress.

\code{mdata_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\details{
This function will wait for all asynchronous operation(s) to
    complete if still in progress (blocking).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

ml <- mirai_walk(
  c(a = 2, b = 3, c = 4), rnorm, mean = 20, .args = list(sd = 2)
)
mdata(ml)

}

}
