% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{mirai_data}
\alias{mirai_data}
\alias{mirai_data_}
\title{mirai (Retrieve Data)}
\usage{
mirai_data(x)

mirai_data_(x)
}
\arguments{
\item{x}{a \sQuote{mirai} object or list of \sQuote{mirai} objects.}
}
\value{
Depending on the type of \sQuote{x} supplied, either an object or a
    list of objects (the same length as \sQuote{x}, preserving names).
}
\description{
\code{mirai_data} retrieves the data value of a \sQuote{mirai}, or list of
    \sQuote{mirai} objects, waiting for resolution if still in progress.

\code{mirai_data_} is a variant that allows user interrupts, suitable for
    interactive use.
}
\details{
This function will wait for the asynchronous operation(s) to
    complete if still in progress (blocking).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

df1 <- data.frame(a = 1, b = 2)
df2 <- data.frame(a = 3, b = 1)
m <- mirai(as.matrix(rbind(df1, df2)), df1 = df1, df2 = df2, .timeout = 1000)
mirai_data(m)

ml <- mwalk(c(a = 2, b = 3, c = 4), rnorm, mean = 20, .args = list(sd = 2))
mirai_data(ml)

}

}
