% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{mirai_version}
\alias{mirai_version}
\title{mirai Version Information}
\usage{
mirai_version()
}
\value{
A character vector of length one.
}
\description{
Provides the package version, \{nanonext\} package version, NNG and mbed TLS
    library versions in a single string.
}
\examples{
mirai_version()

}
