% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mlapply}
\alias{mlapply}
\title{mirai lapply}
\usage{
mlapply(X, FUN, ..., .compute = "default")
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object.}

\item{FUN}{the function to be applied to each element of X.}

\item{...}{optional arguments to \code{FUN}.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
A list (the same length as X, preserving names).
}
\description{
Apply a function over a list or vector using \pkg{mirai}.
}
\details{
This function is blocking and will wait for all results to be
    obtained.
}
\examples{
with(
  daemons(1, dispatcher = FALSE),
  mlapply(1:3, rnorm, mean = 20)
)

}
