% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mwalk}
\alias{mwalk}
\title{mirai Walk}
\usage{
mwalk(.x, .f, ..., .args = list(), .compute = "default")
}
\arguments{
\item{.x}{a list or atomic vector.}

\item{.f}{a function to be applied to each element of \code{.x}.}

\item{...}{optional arguments to \code{.f}.}

\item{.args}{optional arguments to \code{.f} provided as a list.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
An unnamed list of \sQuote{mirai}, the same length as \code{.x}.
}
\description{
Walk a function over a list or vector using \pkg{mirai} for the side-effects.
}
\details{
This function sends each application of \code{.f} on an element of
    \code{.x} for computation in a separate \code{\link{mirai}} call. It does
    not wait for completion.

    Whilst this function is designed primarily to enact side effects, it may
    also be used as an asynchronous map function as \sQuote{mirai} are
    returned, and their values may be collected.

    This function does not perform chunking (assigning multiple elements of
    \code{.x} in batches), and is hence particularly suited to lengthy
    computations or those which have variable or unpredicatable compute times
    over the indices, where it can be optimal to rely on \pkg{mirai}
    scheduling.

    Note: daemons should generally be set prior to calling this function,
    otherwise ephemeral daemons will be created for each computation.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

ml <- mwalk(1:3, rnorm, mean = 20, .args = list(sd = 2))
ml

}

}
