% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfinalpkg-package.R
\docType{package}
\name{myfinalpkg-package}
\alias{myfinalpkg}
\alias{myfinalpkg-package}
\title{myfinalpkg: IntroR4 example pkg}
\description{
Example package for IntroR4 workshop.
}
\keyword{internal}
