% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{nextstream}
\alias{nextstream}
\title{Nextstream}
\usage{
nextstream(.compute = "default")
}
\arguments{
\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
An integer vector of length 7, as given by \code{.Random.seed} when
    the L'Ecuyer-CMRG RNG is in use - this may be passed directly to the 'rs'
    argument of \code{\link{daemon}} in order to set its RNG state, or else
    NULL if daemons have yet to be set (and the stream initialised).
}
\description{
Retrieves the currently stored L'Ecuyer-CMRG RNG stream for the specified
    compute profile and advances it to the next stream.
}
\details{
This function is exported for use by alternative launchers of mirai
    \code{\link{daemon}} processes. The same function is used internally
    within the package by functions that launch daemons.
}
\note{
This function should be called exactly once for its return value. The
    function also has the side effect of automatically advancing the stream
    stored within the compute profile. This ensures that next time the
    function is called the correct value will be returned.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1L)
nextstream()
nextstream()

daemons(0)

}

}
