% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{posit_workbench_config}
\alias{posit_workbench_config}
\title{Posit Workbench Launch Configuration}
\usage{
posit_workbench_config()
}
\value{
A list in the required format to be supplied to the \code{remote} argument
of \code{\link[=daemons]{daemons()}} or \code{\link[=launch_remote]{launch_remote()}}.
}
\description{
Generates a remote configuration for launching daemons via the default
launcher in Posit Workbench. Currently only supports Rstudio Pro sessions.
}
\examples{
tryCatch(posit_workbench_config(), error = identity)

\dontrun{

# Launch 2 daemons using the Posit Workbench default:
daemons(n = 2, url = host_url(), remote = posit_workbench_config()
}

}
\seealso{
\code{\link[=ssh_config]{ssh_config()}}, \code{\link[=cluster_config]{cluster_config()}}, and \code{\link[=remote_config]{remote_config()}} for other
types of remote launch configuration.
}
