% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{registerClusterMirai}
\alias{registerClusterMirai}
\title{Register Cluster Mirai}
\usage{
registerClusterMirai(make.default = TRUE)
}
\arguments{
\item{make.default}{[default TRUE] logical value whether to also register
'miraiCluster' as the default cluster type.}
}
\value{
Invisible NULL
}
\description{
Registers 'miraiCluster' as cluster type 'MIRAI' and optionally makes it the
    default cluster type.
}
\examples{
tryCatch(
  registerClusterMirai(),
  error = as.character
)

}
