% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{saisei}
\alias{saisei}
\title{Saisei - Regenerate Token}
\usage{
saisei(i, .compute = "default")
}
\arguments{
\item{i}{integer \code{i}th daemon to replace.}

\item{.compute}{(optional) character compute profile to use (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
The regenerated character URL upon success, or else NULL.
}
\description{
When using daemons with a local dispatcher service, replaces an existing
    socket at the dispatcher with a new one, listening to a URL incorporating
    a newly-generated token.
}
\details{
Care should be taken when calling this function as the specified
    socket will be closed and replaced immediately without waiting for any
    task in progress to finish.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1, token = TRUE)
daemons()
saisei(i = 1L)
daemons()

daemons(0)

}

}
