% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{scope}
\alias{scope}
\title{Create an Async Scope}
\usage{
scope(x)
}
\arguments{
\item{x}{an expression.}
}
\value{
The return value of the expression \code{x}.
}
\description{
All objects created inside the scope is automatically added to any mirai
calls.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
scope({
  slow_lm <- function(formula, data) {
    Sys.sleep(1)
    lm(formula, data = data)
  }
  dat <- mtcars
  mirai(slow_lm(mpg ~ disp, dat))[]
})
\dontshow{\}) # examplesIf}
}
