% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{sequential_daemons}
\alias{sequential_daemons}
\title{Sequential Daemons}
\usage{
sequential_daemons(.compute = NULL)
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile
to use (each has its own independent set of daemons), or NULL to use the
'default' profile.}
}
\value{
Invisibly, logical \code{TRUE}.
}
\description{
Set daemons which evaluate mirai sequentially in your current R process.
Specify a value for \code{.compute} to limit sequential behaviour to the named
compute profile, otherwise all mirai will be performed sequentially by
default.
}
\details{
Using sequential daemons substantially changes the behaviour of mirai by
evaluating them immediately after mirai creation, hence they are no longer
async. This can, however, be useful for interactive debugging via a
\code{\link[=browser]{browser()}} instance in your current session.

Sequential daemons may be reset in the usual way with a call to \code{daemons(0)}
for the relevant compute profile.
}
\examples{
sequential_daemons()
info()
m <- mirai(Sys.getpid())
daemons(0)
m[]

sequential_daemons("seq")
info("seq")
with_daemons("seq", {
  m <- mirai(Sys.getpid())
})
daemons(0, .compute = "seq")
m[]

}
\seealso{
\code{\link[=daemons]{daemons()}}.
}
