% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{serverq}
\alias{serverq}
\title{mirai Server Queue}
\usage{
serverq(n, .url)
}
\arguments{
\item{n}{integer number of daemons to set.}

\item{.url}{(optional) for distributing tasks across the network: character
client URL and port accepting incoming connections e.g.
'tcp://192.168.0.2:5555' at which server processes started using
\code{\link{server}} should connect to. To listen to port 5555 (for
example) on all interfaces on the host, specify one of 'tcp://:5555',
'tcp://*:5555' or 'tcp://0.0.0.0:5555'.}
}
\value{
Invisible NULL.
}
\description{
Implements an active queue / task scheduler, directing a cluster of daemons.
}
