% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{stop_daemon}
\alias{stop_daemon}
\title{Stop Daemon}
\usage{
stop_daemon(n = 1L, .compute = "default")
}
\arguments{
\item{n}{[default 1L] integer number of daemons to stop.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
Integer number of exit signals sent (the smaller of \sQuote{n} and
  the number of actually connected daemons).
}
\description{
Sends exit signals to reduce the number of connected daemons. Does not
interrupt any \sQuote{mirai} tasks in execution - exits will happen once
these complete (if applicable). This function is only effective when using
dispatcher.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(4)
status()
stop_daemon(2)
mirai(TRUE)[]
status()
daemons(0)

}

}
