% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{workbench_config}
\alias{workbench_config}
\title{Workbench Remote Launch Configuration}
\usage{
workbench_config()
}
\value{
A list in the required format to be supplied to the \code{remote} argument
of \code{\link[=daemons]{daemons()}} or \code{\link[=launch_remote]{launch_remote()}}.
}
\description{
Generates a remote configuration for launching daemons using the default
launcher configured in Posit Workbench.
}
\examples{
tryCatch(workbench_config(), error = identity)

\dontrun{

# Launch 2 daemons using the Workbench default launcher:
daemons(n = 2, url = host_url(), remote = workbench_config())
}

}
\seealso{
\code{\link[=ssh_config]{ssh_config()}}, \code{\link[=cluster_config]{cluster_config()}}, and \code{\link[=remote_config]{remote_config()}} for other
remote launch configurations.
}
