% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-package.R
\name{call_mirai_}
\alias{call_mirai_}
\title{Call mirai}
\usage{
call_mirai_(x)
}
\arguments{
\item{x}{a 'mirai' object, or list of 'mirai' objects.}
}
\description{
\code{call_mirai_} is deprecated and will be removed. Use \code{\link[=call_mirai]{call_mirai()}} instead.
}
\keyword{internal}
