% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{cpinfo}
\alias{cpinfo}
\title{Compute Profile Information}
\usage{
cpinfo(x = c("n", "pid", "tls", "urls"), .compute = "default")
}
\arguments{
\item{x}{[default 'n'] info to request, one of 'n' (integer number of daemons
set or the client URL if not using dispatcher), 'pid' (dispatcher process
ID), 'tls' (TLS configuration), or 'urls' (daemon URLs when using
dispatcher).}

\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
The requested information, or else NULL if not present.
}
\description{
Retrieve information for the selected compute profile (read-only).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1L)
cpinfo("n")
cpinfo("pid")
cpinfo("urls")
daemons(0)

}

}
