% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{daemon}
\alias{daemon}
\title{mirai Server (Async Execution Daemon)}
\usage{
daemon(url)
}
\arguments{
\item{url}{the internally assigned unique URL.}
}
\value{
Integer exit code.
}
\description{
Implements a persistent executor/server for the remote process. Awaits data,
    evaluates an expression in an environment containing the supplied data,
    and returns the result to the caller/client.
}
\keyword{internal}
