% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{daemons_set}
\alias{daemons_set}
\title{Daemons Set}
\usage{
daemons_set(.compute = NULL)
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile to
query, or NULL to query the 'default' profile.

\strong{or} a 'miraiCluster' to obtain its status.}
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical value, whether or not daemons have been set for a given
compute profile.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
daemons_set()
daemons(1)
daemons_set()
daemons(0)
\dontshow{\}) # examplesIf}
}
