% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{daemons_set}
\alias{daemons_set}
\title{Query if Daemons are Set}
\usage{
daemons_set(.compute = NULL)
}
\arguments{
\item{.compute}{character value for the compute profile to use (each has its
own independent set of daemons), or NULL to use the 'default' profile.}
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical value, whether or not daemons have been set for a given
compute profile.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
daemons_set()
daemons(1)
daemons_set()
daemons(0)
\dontshow{\}) # examplesIf}
}
