% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{daemons_set}
\alias{daemons_set}
\title{Query if Daemons are Set}
\usage{
daemons_set(.compute = NULL)
}
\arguments{
\item{.compute}{(character) name of the compute profile. Each profile has its
own independent set of daemons. \code{NULL} (default) uses the 'default'
profile.}
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical value, whether or not daemons have been set for a given
compute profile.
}
\examples{
daemons_set()
daemons(sync = TRUE)
daemons_set()
daemons(0)

}
