% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{daemons_view}
\alias{daemons_view}
\title{View Daemons}
\usage{
daemons_view()
}
\value{
Integer number of daemons.
}
\description{
View the number of currently active 'daemons' or persistent server processes.
}
\examples{
daemons_view()

}
