% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{dispatcher}
\alias{dispatcher}
\title{mirai Dispatcher}
\usage{
dispatcher(
  client,
  url = NULL,
  n = NULL,
  asyncdial = FALSE,
  token = FALSE,
  lock = FALSE,
  refhook = NULL,
  tls = NULL,
  ...,
  monitor = NULL
)
}
\arguments{
\item{client}{the client URL to dial as a character string (where tasks are
sent from), including the port to connect to and (optionally) a path for
websocket URLs e.g. 'tcp://192.168.0.2:5555' or 'ws://192.168.0.2:5555/path'.}

\item{url}{(optional) the URL or range of URLs the dispatcher should
listen at as a character vector, including the port to connect to and
(optionally) a path for websocket URLs e.g. 'tcp://192.168.0.2:5555' or
'ws://192.168.0.2:5555/path'. Tasks are sent to servers dialled into
these URLs. If not supplied, 'n' local inter-process URLs will be
assigned automatically.}

\item{n}{(optional) if specified, the integer number of servers to listen for.
Otherwise 'n' will be inferred from the number of URLs supplied as '...'.
Where a single URL is supplied and 'n' > 1, 'n' unique URLs will be
automatically assigned for servers to dial into.}

\item{asyncdial}{[default FALSE] whether to perform dials asynchronously. The
default FALSE will error if a connection is not immediately possible
(e.g. \code{\link{daemons}} has yet to be called on the client, or the
specified port is not open etc.). Specifying TRUE continues retrying
(indefinitely) if not immediately successful, which is more resilient but
can mask potential connection issues.}

\item{token}{[default FALSE] if TRUE, appends a unique 40-character token
to each URL path the dispatcher listens at (not applicable for TCP URLs
which do not accept a path).}

\item{lock}{[default FALSE] if TRUE, sockets lock once a connection has been
accepted, preventing further connection attempts. This provides safety
against more than one server trying to connect to a unique URL.}

\item{refhook}{[default NULL] maps to the 'refhook' argument of
\code{\link{serialize}} and \code{\link{unserialize}} for providing a
hook function to handle reference objects. As the same function is used
in both cases, it should contain the necessary logic to distinguish
between them as required. Accepts a function or function encoded by
\code{\link[nanonext]{base64enc}}.}

\item{tls}{(optional for secure TLS connections over tls+tcp or wss) maps
directly to the 'server' argument of \code{\link[nanonext]{tls_config}}.
If not provided, a one-time configuration is automatically generated.}

\item{...}{additional arguments passed through to \code{\link{server}} if
launching local daemons i.e. 'url' is not specified.}

\item{monitor}{(for package internal use only) do not set this parameter.}
}
\value{
Invisible NULL.
}
\description{
Implements a dispatcher for tasks from a client to multiple servers for
    processing, using a FIFO scheduling rule, queuing tasks as required.
}
\details{
The network topology is such that a dispatcher acts as a gateway
    between clients and servers, ensuring that tasks received from clients
    are dispatched on a FIFO basis to servers for processing. Tasks are
    queued at the dispatcher to ensure tasks are only sent to servers that
    can begin immediate execution of the task.
}
