% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{.}
\alias{.}
\title{mirai Server (Async Execution Daemon)}
\usage{
.(.)
}
\arguments{
\item{.}{the client URL specifying the transport and address as a character
string e.g. 'tcp://192.168.0.2:5555'}
}
\value{
Integer exit code.
}
\description{
Implements a persistent executor/server for the remote process. Awaits data,
    evaluates an expression in an environment containing the supplied data,
    and returns the result to the caller/client.
}
