% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{.}
\alias{.}
\title{mirai Server (Async Executor [Daemon])}
\usage{
.(.url, .. = TRUE)
}
\arguments{
\item{.url}{the client URL and port to connect to as a character string e.g.
'tcp://192.168.0.2:5555'.}

\item{..}{[default TRUE] launch as a persistent daemon or, if FALSE, an
ephemeral process.}
}
\value{
Invisible NULL.
}
\description{
Implements a [persistent] executor/server for the remote process. Awaits data,
    evaluates an expression in an environment containing the supplied data,
    and returns the result to the caller/client.
}
