% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{.daemon}
\alias{.daemon}
\title{dot Daemon}
\usage{
.daemon(url)
}
\arguments{
\item{url}{the character host or dispatcher URL to dial into, including the
port to connect to (and optionally for websockets, a path), e.g.
'tcp://hostname:5555' or 'ws://10.75.32.70:5555/path'.}
}
\value{
Logical TRUE or FALSE.
}
\description{
Ephemeral executor for the remote process. User code must not call this.
    Consider \code{daemon(maxtasks = 1L)} instead.
}
\keyword{internal}
