% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{.daemon}
\alias{.daemon}
\title{dot Daemon}
\usage{
.daemon(url)
}
\arguments{
\item{url}{the character host or dispatcher URL to dial into, including the
port to connect to (and optionally for websockets, a path), e.g.
'tcp://192.168.0.2:5555' or 'ws://192.168.0.2:5555/path'.}
}
\value{
Logical TRUE, invisibly.
}
\description{
Implements an ephemeral executor for the remote process.
}
\keyword{internal}
