% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{.flat}
\alias{.flat}
\alias{.progress}
\alias{.stop}
\title{mirai Map Options}
\format{
An object of class \code{expression} of length 1.

An object of class \code{expression} of length 1.

An object of class \code{expression} of length 1.
}
\usage{
.flat

.progress

.stop
}
\description{
Expressions to insert into the \code{[]} method for \sQuote{mirai_map}
    objects.
}
\section{Results}{


    \code{x[]} collects the results of a mirai_map \code{x}. This will wait
    for all asynchronous operations to complete if still in progress,
    blocking but user-interruptible.

    \code{x[.flat]} collects and flattens map results, avoiding coercion.
    Note: errors if an \sQuote{errorValue} has been returned or results are
    of differing type.

    \code{x[.progress]} collects map results whilst showing a text progress
    indicator.

    \code{x[.stop]} collects map results applying early stopping, which stops
    at the first failure and aborts all remaining queued operations. Note:
    individual operations already in progress continue to completion,
    although their results are not collected.

    The options above may be combined in a vector, for example: \cr
    \code{x[c(.stop, .progress)]} applies early stopping together with a
    progress indicator.
}

\keyword{datasets}
\keyword{internal}
