% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{.flat}
\alias{.flat}
\alias{.progress}
\alias{.progress_cli}
\alias{.stop}
\title{mirai Map Options}
\format{
An object of class \code{bytecode} of length 1.

An object of class \code{bytecode} of length 1.

An object of class \code{bytecode} of length 1.

An object of class \code{bytecode} of length 1.
}
\usage{
.flat

.progress

.progress_cli

.stop
}
\description{
Expressions to be provided to the \code{[]} method for \sQuote{mirai_map}
objects, or the \code{...} argument of \code{\link{collect_mirai}}.
}
\section{Collection Options}{


\code{x[]} collects the results of a \sQuote{mirai_map} \code{x} and returns
a list. This will wait for all asynchronous operations to complete if still
in progress, blocking but user-interruptible.

\code{x[.flat]} collects and flattens map results to a vector, checking that
they are of the same type to avoid coercion. Note: errors if an
\sQuote{errorValue} has been returned or results are of differing type.

\code{x[.progress]} collects map results whilst showing a simple text
progress indicator of parts completed of the total.

\code{x[.progress_cli]} collects map results whilst showing a progress bar
from the \CRANpkg{cli} package, if available, with completion percentage and
ETA.

\code{x[.stop]} collects map results applying early stopping, which stops at
the first failure and cancels remaining operations. Note: operations already
in progress continue to completion, although their results are not collected.

The options above may be combined in the manner of: \cr
\code{x[.stop, .progress]} which applies early stopping together with a
progress indicator.
}

\keyword{datasets}
\keyword{internal}
