% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{everywhere}
\alias{everywhere}
\title{Evaluate Everywhere}
\usage{
everywhere(.expr, ..., .args = list(), .compute = "default")
}
\arguments{
\item{.expr}{an expression to evaluate asynchronously (of arbitrary length,
wrapped in \{\} if necessary), \strong{or} a language object passed by
\link{name}.}

\item{...}{(optional) named arguments (name = value pairs) specifying
objects referenced in '.expr'. Used in addition to, and taking precedence
over, any arguments specified via '.args'.}

\item{.args}{(optional) \strong{either} a list of objects to be passed by
\link{name} (found in the current scope), \strong{or else} a list of
name = value pairs, as in '...'.}

\item{.compute}{[default 'default'] character value for the compute profile
to use when sending the mirai.}
}
\value{
Invisible NULL.
}
\description{
Evaluate an expression 'everywhere' on all connected daemons for the
    specified compute profile. Designed for performing setup operations
    across daemons, resultant changes to the global environment, loaded
    packages or options are persisted regardless of the daemon's 'cleanup'
    setting.
}
\note{
When not using dispatcher: daemons are launched asynchronously, and
    require a short time delay between launch and calling this function.
    Alternatively, supply an integer value for 'seed' as part of the '...'
    argument to \code{\link{daemons}}, which effects a synchronous launch,
    making it safe to call this function immediately afterwards.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1)
everywhere(list2env(x, envir = .GlobalEnv), x = list(a = 1, b = 2))
m <- mirai(a + b)
call_mirai(m)$data
daemons(0)

daemons(1, dispatcher = FALSE, seed = 123)
everywhere(library(parallel))
m <- mirai(setDefaultCluster())
call_mirai(m)$data
daemons(0)

}

}
