% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{exec}
\alias{exec}
\title{mirai Server (Async Execution Daemon)}
\usage{
exec(expr, url)
}
\arguments{
\item{expr}{an expression to evaluate in a new R process.}
}
\value{
Invisible NULL.
}
\description{
Implements an executor/server for the remote process. Awaits data, evaluates
    an expression in an environment containing the supplied data, and returns
    the result to the caller/client.
}
\keyword{internal}
