% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{info}
\alias{info}
\title{Information Statistics}
\usage{
info(.compute = "default")
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile to
query, or NULL to query the 'default' profile.

\strong{or} a 'miraiCluster' to obtain its status.}
}
\value{
Integer vector with the names: connections, cumulative, awaiting,
executing, completed. Or else \code{NULL} if the compute profile is yet to be
set up.
}
\description{
Retrieve statistics for the specified compute profile.
}
