% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{info}
\alias{info}
\title{Information Statistics}
\usage{
info(.compute = "default")
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile to
query, or NULL to query the 'default' profile.

\strong{or} a 'miraiCluster' to obtain its status.}
}
\value{
Named integer vector or else \code{NULL} if the compute profile is yet to
be set up. The names comprise: 'connections', 'cumulative', 'awaiting',
'executing', and 'completed'. For non-dispatcher daemons: only
'connections' will be available and the other values will be \code{NA}.
}
\description{
Retrieve statistics for the specified compute profile.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
info()
daemons(1)
info()
daemons(0)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=status]{status()}} for more verbose status information.
}
