% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_error_value}
\alias{is_error_value}
\title{Is Error Value}
\usage{
is_error_value(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE if 'x' is of class 'errorValue', FALSE otherwise.
}
\description{
Is the object an error value, such as a 'mirai' timeout, a 'miraiError' from
    failed execution within a 'mirai' or a 'miraiInterrupt' resulting from
    the user interrupt of an ongoing 'mirai' evaluation.
}
\examples{
is_error_value(1L)

}
