% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_mirai}
\alias{is_mirai}
\title{Is mirai}
\usage{
is_mirai(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical TRUE if \sQuote{x} is of class \sQuote{mirai}, FALSE
    otherwise.
}
\description{
Is the object a \sQuote{mirai}.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

df <- data.frame()
m <- mirai(as.matrix(df), df = df)
is_mirai(m)
is_mirai(df)

}

}
