% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_mirai_error}
\alias{is_mirai_error}
\title{Is mirai_error}
\usage{
is_mirai_error(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical value TRUE or FALSE.
}
\description{
Is the object a 'mirai_error'. When execution fails in a mirai process, the
    error message is returned as a character string classed as 'mirai_error'.
    To test for timeouts, \code{\link{is_error_value}} should be used instead.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

m <- mirai(stop())
call_mirai(m)
is_mirai_error(m$data)

}

}
