% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{is_mirai_error}
\alias{is_mirai_error}
\title{Is mirai Error}
\usage{
is_mirai_error(x)
}
\arguments{
\item{x}{an object.}
}
\value{
Logical TRUE if 'x' is of class 'miraiError', FALSE otherwise.
}
\description{
Is the object a 'miraiError'. When execution in a mirai process fails, the
    error message is returned as a character string of class 'miraiError' and
    'errorValue'. To test for all error conditions, including timeouts etc.,
    \code{\link{is_error_value}} should be used instead.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

m <- mirai(stop())
call_mirai(m)
is_mirai_error(m$data)

}

}
