% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{launch}
\alias{launch}
\title{Launch Background Process}
\usage{
launch(args)
}
\arguments{
\item{args}{character string. This will be shell quoted by \code{\link{shQuote}}.}
}
\value{
Invisibly, an exit code (zero upon success).
}
\description{
Uses \code{\link{system2}} to call \code{Rscript} in a background process,
    with 'args' passed as command line argument to \code{Rscript -e 'args'}.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

launch('mirai::server("abstract://mirai", idletime = 60000L)')

}

}
