% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{make_cluster}
\alias{make_cluster}
\alias{stop_cluster}
\title{Make Mirai Cluster}
\usage{
make_cluster(
  n,
  url = NULL,
  ...,
  command = NULL,
  args = c("", "."),
  rscript = "Rscript"
)

stop_cluster(cl)
}
\arguments{
\item{n}{integer number of nodes (launched on the local machine unless 'url'
is specified).}

\item{url}{(specify for remote nodes), the character URL on the host for
remote nodes to dial into, including a port accepting incoming connections,
e.g. 'tcp://10.75.37.40:5555'. Specify a URL starting 'tls+tcp://' to use
secure TLS connections.}

\item{...}{additional arguments passed onto \code{\link{daemons}}.}

\item{command}{(optional) the command used to effect the daemon launch on the
remote machine as a character string (e.g. \code{'ssh'}).}

\item{args}{(optional) arguments passed to 'command', as a character vector
that must include \code{"."} as an element, which will be substituted
for the daemon launch command. Alternatively, a list of character
vectors, such as those produced by \code{\link{ssh_args}}, a convenience
function for constructing valid SSH arguments.}

\item{rscript}{(for remote only) name / path of the Rscript executable on
the remote machine. The default assumes 'Rscript' is on the executable
search path. Prepend the full path if necessary. If launching on Windows,
'Rscript' should be replaced with 'Rscript.exe'.}

\item{cl}{a 'miraiCluster'.}
}
\value{
For \strong{make_cluster}: An object of class 'miraiCluster' and
    'cluster'. Each 'miraiCluster' has an automatically assigned ID and 'n'
    nodes of class 'miraiNode'.

    For \strong{stop_cluster}: invisible NULL.
}
\description{
\code{make_cluster} creates a cluster of type 'miraiCluster', which may be
    used as a cluster object for any function in the \pkg{parallel} base
    package such as \code{\link[parallel]{clusterApply}} or
    \code{\link[parallel]{parLapply}}.

\code{stop_cluster} stops a cluster created by \code{make_cluster}.
}
\details{
The default behaviour of clusters created by this function is
    designed to map as closely as possible to clusters created by the
    \pkg{parallel} package. However, '...' arguments are passed onto
    \code{\link{daemons}} for additional customisation if desired.

    Call \code{\link{status}} on a 'miraiCluster' to check the number of
    currently active connections as well as the host URL.
}
\note{
Requires R >= 4.4 (currently R-devel). Clusters created with this
    function will not work with prior R versions. The functionality is
    experimental prior to release of R 4.4 and the interface is consequently
    subject to change at any time.
}
\section{Remote Nodes}{


    For remote nodes, the arguments 'command', 'args' and 'rscript' map
    directly to the relevant arguments of \code{\link{launch_remote}}.

    If launching remote nodes, the number of nodes is inferred from the length
    of 'args', and 'n' is disregarded if supplied. The auxiliary function
    \code{\link{ssh_args}} may be used to construct the required arguments to
    use SSH as a launcher, both with and without tunnelling.

    Alternatively, by specifying only 'url' and 'n', the host connection
    listening at 'url' is set up, and nodes may be launched by other custom
    means.
}

\examples{
if (interactive()) {
# Only run examples in interactive R sessions

cl <- make_cluster(2)
cl
cl[[1L]]

Sys.sleep(0.5)
status(cl)

stop_cluster(cl)

}

}
