% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-package.R
\docType{package}
\name{mirai-package}
\alias{mirai-package}
\title{mirai: Minimalist Async Evaluation Framework for R}
\description{
Designed for simplicity, a 'mirai' evaluates an R expression asynchronously
in a parallel process, locally or distributed over the network. Modern
networking and concurrency, built on 'nanonext' and 'NNG', ensures reliable
scheduling over fast inter-process communications or TCP/IP secured by TLS.
Launch remote resources via SSH or cluster managers for distributed
computing. Scales efficiently to millions of tasks over thousands of
connections, requiring no storage on the file system due to its inherently
queued architecture. Innovative features include event-driven promises,
asynchronous parallel map, and seamless serialization of otherwise
non-exportable reference objects.
}
\section{Notes}{


For local mirai requests, the default transport for inter-process
communications is platform-dependent: abstract Unix domain sockets on Linux,
Unix domain sockets on MacOS, Solaris and other POSIX platforms, and named
pipes on Windows.

This may be overriden, if desired, by specifying 'url' in the \code{\link[=daemons]{daemons()}}
interface and launching daemons using \code{\link[=launch_local]{launch_local()}}.
}

\section{Reference Manual}{


\code{vignette("mirai", package = "mirai")}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mirai.r-lib.org}
  \item \url{https://github.com/r-lib/mirai}
  \item Report bugs at \url{https://github.com/r-lib/mirai/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Joe Cheng \email{joe@posit.co} [contributor]
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
  \item Hibiki AI Limited [copyright holder]
}

}
