% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-package.R
\docType{package}
\encoding{UTF-8}
\name{mirai-package}
\alias{mirai-package}
\title{mirai: Minimalist Async Evaluation Framework for R}
\description{
Simple and lightweight parallel code execution, local or distributed across
    the network, built on 'nanonext' and 'NNG' (Nanomsg Next Gen) technology.
    A 'mirai' (Japanese for 'future') evaluates an arbitrary expression
    asynchronously, resolving automatically upon completion.
}
\section{Implementation}{


    'mirai' also inherit from the class 'recvAio' and are asynchronous
    message receive objects using the \{nanonext\} framework.

    Functions \code{\link{call_mirai}}, \code{\link{stop_mirai}} and
    \code{\link{unresolved}} are directly assigned from 'nanonext' for
    efficiency. As a result, their argument name remains 'aio' instead of
    'mirai'. This should be of no practical significance as they all take a
    single argument, hence normal usage precludes use of the argument name.
}

\section{Links}{


    mirai website: \url{https://shikokuchuo.net/mirai/}\cr
    mirai on CRAN: \url{https://cran.r-project.org/package=mirai}

    nanonext website: \url{https://shikokuchuo.net/nanonext/}\cr
    nanonext on CRAN: \url{https://cran.r-project.org/package=nanonext}

    NNG website: \url{https://nng.nanomsg.org/}
}

\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
    (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
