% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-package.R
\docType{package}
\encoding{UTF-8}
\name{mirai-package}
\alias{mirai-package}
\title{mirai: Minimalist Async Evaluation Framework for R}
\description{
Extremely simple and lightweight method for concurrent / parallel code
    execution, built on 'nanonext' and 'NNG' (Nanomsg Next Gen) technology.
    mirai is Japanese for 'future'.
}
\section{Links}{


    mirai website: \url{https://shikokuchuo.net/mirai/}\cr
    mirai on CRAN: \url{https://cran.r-project.org/package=mirai}

    nanonext website: \url{https://shikokuchuo.net/nanonext/}\cr
    nanonext on CRAN: \url{https://cran.r-project.org/package=nanonext}

    NNG website: \url{https://nng.nanomsg.org/}
}

\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
    (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
