% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{mirai}
\alias{mirai}
\title{mirai (Control Panel)}
\usage{
mirai(..., set_daemons, view_daemons)
}
\arguments{
\item{...}{reserved.}

\item{set_daemons}{set an integer number of background processes.}

\item{view_daemons}{specify any value to return the current number of
background processes.}
}
\value{
Depending on specified parameters:
    \itemize{
    \item{'set_daemons': the return value will depend on whether background
    processes are created or destroyed (see Daemons section).}
    \item{'view_daemons': the integer number of background processes.}
    \item{none: the 'nanoSocket' for connecting to the background processes,
    or NULL if it has yet to be created.}
    }
}
\description{
Settings that controls the overall behaviour of \{mirai\}. Note that as all
    arguments appear after '...', they must be specified explicitly and in
    full (positional and partial matching do not apply). Specify a single
    argument only: if more than one is specified, only the first will be
    taken into account, in the order they appear in the argument list below.
}
\section{Daemons}{


    Set or view the number of daemons (background processes). Create
    persistent background processes to send \code{\link{eval_mirai}} requests.
    Setting a positive number of daemons provides a potentially more efficient
    solution for async operations as new processes do not need to be created
    on an ad hoc basis. [Experimental]

    Background processes will be created or destroyed as appropriate.
    \itemize{
    \item{If new processes are created, the return value will be the integer
    number of created processes.}
    \item{If processes are destroyed, the return value will be a list of the
    exit signals from each destroyed process (an integer byte 1 on success).
    This process will wait for confirmation to be received.}
    \item{Otherwise NULL will be returned invisibly.}
    }

    mirai will revert to the default behaviour of creating a new background
    process for each request if the number of daemons is set to 0.

    This feature has the tag [experimental], which indicates that it remains
    under development. Please note that the final implementation may differ
    from the current version.
}

\examples{
if (interactive()) {
# Only run examples in interactive R sessions

# To create 4 background processes
mirai(set_daemons = 4)
# To view the number of background processes
mirai(view_daemons = TRUE)
# To destroy them all
mirai(set_daemons = 0)
}

}
