% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{mirai}
\alias{mirai}
\title{mirai (Daemon Manager)}
\usage{
mirai(set_daemons)
}
\arguments{
\item{set_daemons}{integer number of background processes.}
}
\value{
The return value will depend on whether background processes are
    created or destroyed (see details section). Without specifying any
    arguments, the 'nanoSocket' for connecting to the background
    processes, or NULL if it has yet to be created.
}
\description{
Set the number of daemons (background processes). Use this function
    to create persistent background processes to send \code{\link{eval_mirai}}
    requests. Setting a positive number of daemons provides a potentially
    more efficient solution for async operations as new processes do not need
    to be spun up on an ad hoc basis. [Experimental]
}
\details{
Background processes will be created or destroyed as appropriate.
    \itemize{
    \item{If new processes are created, the return value will be the integer
    number of created processes.}
    \item{If processes are destroyed, the return value will be a list of the
    exit signals from each destroyed process (an integer byte 1 on success).
    This process will wait for confirmation to be received.}
    \item{Otherwise NULL will be returned invisibly.}
    }

    mirai will revert to the default behaviour of creating a new background
    process for each request if the number of daemons is set to 0.

    This feature has the tag [experimental], which indicates that it remains
    under development. Please note that the final implementation may differ
    from the current version.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions
# To spin up 4 background processes
mirai(4)
# To destroy them all
mirai(0)
}

}
