% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mmap}
\alias{mmap}
\title{mirai Map}
\usage{
mmap(.x, .f, ..., .args = list(), .compute = "default")
}
\arguments{
\item{.x}{a list, atomic vector, or an expression object.}

\item{.f}{a function to be applied to each element of \code{.x}.}

\item{...}{optional arguments to \code{.f}.}

\item{.args}{optional arguments to \code{.f} provided as a list.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
A list (the same length as \code{.x}, preserving names).
}
\description{
Map a function over a list or vector using \pkg{mirai}.
}
\details{
This function is blocking and will wait for all results to be
    obtained. Daemons must also have been previously set with a call to
    \code{\link{daemons}}.
}
\examples{
with(
  daemons(1, dispatcher = FALSE),
  mmap(1:3, rnorm, mean = 20, .args = list(sd = 2))
)

}
