% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{mmap}
\alias{mmap}
\title{mirai Map}
\usage{
mmap(
  .x,
  .f,
  ...,
  .args = list(),
  .progress = FALSE,
  .stop = FALSE,
  .compute = "default"
)
}
\arguments{
\item{.x}{a list or atomic vector.}

\item{.f}{a function to be applied to each element of \code{.x}.}

\item{...}{optional arguments to \code{.f}.}

\item{.args}{optional arguments to \code{.f} provided as a list.}

\item{.progress}{[default FALSE] if TRUE, reports progress via a simple text
progress indicator to \sQuote{stderr}.}

\item{.stop}{[default FALSE] errors are returned as \sQuote{miraiError} /
\sQuote{errorValue} as the case may be, allowing recovery from partial
failure. If TRUE, performs early stopping as soon as an error is
encountered, with remaining computations aborted.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
A list (the same length as \code{.x}, preserving names).
}
\description{
Map a function over a list or vector using \pkg{mirai}.
}
\details{
This function sends each application of \code{.f} on an element of
    \code{.x} for computation in a separate \code{\link{mirai}} call, and
    waits for completion.

    Designed to facilitate recovery from partial failure by returning all
    results by default, allowing only the failures to be re-run.

    Alternatively, there is the option for early stopping, which stops at the
    first failure and aborts all remaining computations.

    Daemons should have previously been set with a call to
    \code{\link{daemons}}; if not then operations will be performed in a
    single ephemeral daemon and a warning is output.
}
\examples{
with(
  daemons(1, dispatcher = FALSE),
  mmap(1:3, rnorm, mean = 20, .args = list(sd = 2))
)

if (interactive()) {

mmap(seq(from = 0.1, to = 0.4, by = 0.1), Sys.sleep, .progress = TRUE)

}

}
