% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{on_daemon}
\alias{on_daemon}
\title{On a daemon}
\usage{
on_daemon()
}
\value{
Logical \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical value, whether evaluation is taking place within a mirai
call on a daemon.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
on_daemon()
mirai(mirai::on_daemon())[]
\dontshow{\}) # examplesIf}
}
