% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{register}
\alias{register}
\title{Register Custom Serialization Functions}
\usage{
register(refhook = list())
}
\arguments{
\item{refhook}{\strong{either} a list of two functions: the signature for the
first must accept a list of external pointer type objects and return a
raw vector, e.g. \code{torch::torch_serialize}, and the second must
accept a raw vector and return a list of external pointer type objects,
e.g. \code{torch::torch_load},\cr \strong{or else} NULL to reset.}
}
\value{
A list comprising the currently-registered 'refhook' functions.
}
\description{
Registers custom serialization and unserialization functions for sending and
    receiving reference objects accessed via an external pointer.
}
\details{
Calling without any arguments returns a list of the
    currently-registered 'refhook' functions.

    This function may be called prior to or after setting daemons, with the
    registered functions applying across all compute profiles.
}
\examples{
r <- register(list(function(x) serialize(x, NULL), unserialize))
register()
register(r)

register(NULL)
register()

}
