% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{register_cluster}
\alias{register_cluster}
\title{Register Mirai Cluster}
\usage{
register_cluster(default = TRUE)
}
\arguments{
\item{default}{[default TRUE] logical value whether to also register
'miraiCluster' as the default cluster type.}
}
\value{
Invisible NULL
}
\description{
Registers 'miraiCluster' with the \pkg{parallel} package as cluster type
    'MIRAI' and (optionally) makes it the default cluster type.
}
\examples{
tryCatch(

mirai:::register_cluster()

, error = identity)

}
\keyword{internal}
