% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{remote_config}
\alias{remote_config}
\alias{ssh_config}
\title{Generic and SSH Remote Launch Configuration}
\usage{
remote_config(
  command = NULL,
  args = c("", "."),
  rscript = "Rscript",
  quote = FALSE
)

ssh_config(
  remotes,
  tunnel = FALSE,
  timeout = 10,
  command = "ssh",
  rscript = "Rscript",
  host
)
}
\arguments{
\item{command}{the command used to effect the daemon launch on the remote
machine as a character string (e.g. \code{'ssh'}). Defaults to
\sQuote{ssh} for \code{ssh_config}, although may be substituted for the
full path to a specific SSH application. The default NULL for
\code{remote_config} does not effect any launches, but causes
\code{\link{launch_remote}} to return the shell commands for manual
deployment on remote machines.}

\item{args}{(optional) arguments passed to \sQuote{command}, as a character
vector that must include \code{"."} as an element, which will be
substituted for the daemon launch command. Alternatively, a list of such
character vectors to effect multiple launches (one for each list element).}

\item{rscript}{(optional) name / path of the Rscript executable on the remote
machine. The default assumes \sQuote{Rscript} is on the executable search
path. Prepend the full path if necessary. If launching on Windows,
\sQuote{Rscript} should be replaced with \sQuote{Rscript.exe}.}

\item{quote}{[default FALSE] logical value whether or not to quote the daemon
launch command (not required for SLURM \sQuote{srun} for example, but
required for \sQuote{ssh}).}

\item{remotes}{the character URL or vector of URLs to SSH into, using the
'ssh://' scheme and including the port open for SSH connections (defaults
to 22 if not specified), e.g. 'ssh://10.75.32.90:22' or 'ssh://nodename'.}

\item{tunnel}{[default FALSE] logical value whether to use SSH reverse
tunnelling. If TRUE, a tunnel is created between the same ports on the
local and remote machines. See the \sQuote{SSH Tunnelling} section below
for how to correctly specify required settings.}

\item{timeout}{[default 10] maximum time allowed for connection setup in
seconds.}

\item{host}{(optional) only applicable for reverse tunnelling. Should be
specified if creating a standalone configuration object. If calling this
function directly as an argument to \code{\link{daemons}}, this is not
required and can be inferred from the \sQuote{url} supplied (see
\sQuote{SSH Tunnelling} section below).}
}
\value{
A list in the required format to be supplied to the \sQuote{remote}
    argument of \code{\link{launch_remote}}, \code{\link{daemons}}, or
    \code{\link{make_cluster}}.
}
\description{
\code{remote_config} provides a flexible generic framework for generating
    the shell commands to deploy daemons remotely.

\code{ssh_config} generates a remote configuration for launching daemons over
    SSH, with the option of SSH tunnelling.
}
\section{SSH Direct Connections}{


    The simplest use of SSH is to execute the daemon launch command on a
    remote machine, for it to dial back to the host / dispatcher URL.

    It is assumed that SSH key-based authentication is already in place. The
    relevant port on the host must also be open to inbound connections from
    the remote machine.
}

\section{SSH Tunnelling}{


    Use of SSH tunnelling provides a convenient way to launch remote nodes
    without requiring the remote machine to be able to access the host.
    Often firewall configurations or security policies may prevent opening a
    port to accept outside connections.

    In these cases SSH tunnelling offers a solution by creating a tunnel once
    the initial SSH connection is made. For simplicity, this SSH tunnelling
    implementation uses the same port on both the side of the host and that
    of the corresponding node. SSH key-based authentication must also already
    be in place.

    Tunnelling requires the hostname for the \sQuote{host} argument (or the
    \sQuote{url} argument to \code{\link{daemons}} if called directly in
    that context) to be either \sQuote{127.0.0.1} or \sQuote{localhost}.
    This is as the tunnel is created between \code{127.0.0.1:port} or
    equivalently \code{localhost:port} on each machine. The host listens to
    \code{port} on its machine and the remotes each dial into \code{port} on
    their own respective machines.
}

\examples{
# for SLURM
remote_config(
  command = "srun",
  args = c("--mem 512", "-n 1", "."),
  rscript = file.path(R.home("bin"), "Rscript")
)

# commands like SSH require quoting of the daemon launch command
remote_config(
  command = "/usr/bin/ssh",
  args = c("-fTp 22 10.75.32.90", "."),
  quote = TRUE
)

ssh_config(
  remotes = c("ssh://10.75.32.90:222", "ssh://nodename"),
  timeout = 5
)

ssh_config(
  remotes = c("ssh://10.75.32.90:222", "ssh://nodename"),
  tunnel = TRUE,
  host = "tls+tcp://127.0.0.1:5555"
)

\dontrun{

# launch 2 daemons on the remote machines 10.75.32.90 and 10.75.32.91 using
# SSH, connecting back directly to the host URL over a TLS connection:

daemons(
  url = host_url(tls = TRUE),
  remote = ssh_config(
    remotes = c("ssh://10.75.32.90:222", "ssh://10.75.32.91:222"),
    timeout = 1
  )
)

# launch 2 nodes on the remote machine 10.75.32.90 using SSH tunnelling over
# port 5555 ('url' hostname must be 'localhost' or '127.0.0.1'):

cl <- make_cluster(
  url = "tcp://localhost:5555",
  remote = ssh_config(
    remotes = c("ssh://10.75.32.90", "ssh://10.75.32.90"),
    tunnel = TRUE,
    timeout = 1
  )
)
}

}
