% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{require_daemons}
\alias{require_daemons}
\title{Require Daemons}
\usage{
require_daemons(.compute = NULL, call = environment())
}
\arguments{
\item{.compute}{(character) name of the compute profile. Each profile has its
own independent set of daemons. \code{NULL} (default) uses the 'default'
profile.}

\item{call}{(environment) execution environment for error attribution, e.g.
\code{environment()}. Used by \pkg{cli} for error messages.}
}
\value{
Invisibly, logical \code{TRUE}, or else errors.
}
\description{
Returns \code{TRUE} invisibly only if daemons are set, otherwise produces an
informative error for the user to set daemons, with a clickable function link
if the \pkg{cli} package is available.
}
\examples{
daemons(sync = TRUE)
(require_daemons())
daemons(0)

}
