% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{require_daemons}
\alias{require_daemons}
\title{Require Daemons}
\usage{
require_daemons(.compute = NULL, call = environment())
}
\arguments{
\item{.compute}{character value for the compute profile to use (each has its
own independent set of daemons), or NULL to use the 'default' profile.}

\item{call}{(only used if the \CRANpkg{cli} package is installed) the
execution environment of a currently running function, e.g.
\code{environment()}. The function will be mentioned in error messages as the
source of the error.}
}
\value{
Invisibly, logical \code{TRUE}, or else errors.
}
\description{
Returns \code{TRUE} invisibly only if daemons are set, otherwise produces an
informative error for the user to set daemons, with a clickable function link
if the \pkg{cli} package is available.
}
\examples{
daemons(sync = TRUE)
(require_daemons())
daemons(0)

}
