% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{saisei}
\alias{saisei}
\title{Saisei - Regenerate Token}
\usage{
saisei(i, force = FALSE, .compute = "default")
}
\arguments{
\item{i}{integer \code{i}th URL to replace.}

\item{force}{[default FALSE] logical value whether to replace the listener
even when there is an existing connection.}

\item{.compute}{[default 'default'] character compute profile to use (each
compute profile has its own set of daemons for connecting to different
resources).}
}
\value{
The regenerated character URL upon success, or else invisible NULL.

    If \code{saisei()} is called with no arguments, a named list of two
    elements: \code{$urls} - a character vector of the URLs dispatcher is
    listening at, and \code{$pid} - the integer process ID for dispatcher.
}
\description{
When using daemons with a local dispatcher service, regenerates the token for
    the URL a dispatcher socket listens at.
}
\details{
As the specified listener is closed and replaced immediately, this
    function will only be successful if there are no existing connections at
    the socket (i.e. 'online' status shows 0), unless the argument 'force' is
    specified as TRUE.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

daemons(1L)
Sys.sleep(1L)
daemons()
saisei(i = 1L, force = TRUE)
daemons()
saisei()

daemons(0)

}

}
