% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{saisei}
\alias{saisei}
\title{Saisei (Regenerate Token)}
\usage{
saisei(i, force = FALSE, .compute = "default")
}
\arguments{
\item{i}{integer index number URL to regenerate at dispatcher.}

\item{force}{[default FALSE] logical value whether to regenerate the URL even
when there is an existing active connection.}

\item{.compute}{[default 'default'] character value for the compute profile
to use (each compute profile has its own independent set of daemons).}
}
\value{
The regenerated character URL upon success, or else NULL.
}
\description{
[DEPRECATED] This is a legacy function used only with the legacy v1
dispatcher, and will be removed in a future release.
}
\keyword{internal}
