% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{server}
\alias{server}
\title{mirai Server (Async Executor Daemon)}
\usage{
server(
  url,
  nodes = NULL,
  baseurl = NULL,
  idletime = Inf,
  runtime = Inf,
  tasks = Inf
)
}
\arguments{
\item{url}{the client URL and port to connect to as a character string e.g.
'tcp://192.168.0.2:5555'.}

\item{nodes}{[default NULL] if supplied, this server instance will run as an
active queue (task scheduler) with the specified number of nodes.}

\item{baseurl}{[default NULL] if supplied (together with 'nodes'), the client
URL for nodes to connect to with a starting port number e.g.
'tcp://192.168.0.2:5555' with 6 nodes uses the contiguous block of ports
5555 through 5560. Otherwise, nodes are automatically launched on the
same machine.}

\item{idletime}{[default Inf] maximum idle time, since completion of the last
task (in milliseconds) before exiting.}

\item{runtime}{[default Inf] soft walltime, or the minimum amount of real
time taken (in milliseconds) before exiting.}

\item{tasks}{[default Inf] the maximum number of tasks to execute (task
limit) before exiting.}
}
\value{
Invisible NULL.
}
\description{
Implements a persistent executor/server for the remote process. Awaits data,
    evaluates an expression in an environment containing the supplied data,
    and returns the result to the caller/client.
}
\details{
The network topology is such that server daemons dial into the
    client, which listens at the '.url' address. In this way, network
    resources may be easily added or removed at any time and the client
    automatically distributes tasks to all available servers.

    If 'nodes' is supplied, this daemon is launched as an active server queue,
    directing a cluster with the specified number of nodes. A server queue
    may be used in combination with other servers or server queues.

    If 'baseurl' is also supplied, this active server queue allocates and
    listens to a block of URLs with ports starting from the supplied port
    number and incrementing by one. Individual \code{\link{server}} instances
    should then be started on the remote resource, with these as the client
    'url'.

    If 'nodes' is supplied without 'baseurl' then the 'nodes' are launched
    automatically by the active server queue on the same machine.
}
