% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{serverq}
\alias{serverq}
\title{mirai Server Queue}
\usage{
serverq(n, .url)
}
\arguments{
\item{n}{integer number of daemons to set.}

\item{.url}{the client URL and port to connect to as a character string e.g.
'tcp://192.168.0.2:5555'.}
}
\value{
Invisible NULL.
}
\description{
Implements an active queue / task scheduler, launching and directing a
    cluster of daemons.
}
\section{About}{


    The network topology is such that this server queue dials into the client,
    which listens at the '.url' address. A server queue launches a cluster of
    'n' daemons, and relays messages back and forth from the client. A server
    queue may be used in combination with other servers or server queues and
    the client automatically distributes tasks to all available resources.
}

