% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{status}
\alias{status}
\title{Status Information}
\usage{
status(.compute = NULL)
}
\arguments{
\item{.compute}{(character | miraiCluster) compute profile name, or \code{NULL}
for 'default'. Also accepts a 'miraiCluster'.}
}
\value{
A named list comprising:
\itemize{
\item \strong{connections} - integer number of active daemon connections.
\item \strong{daemons} - character URL at which host / dispatcher is
listening, or else \code{0L} if daemons have not yet been set.
\item \strong{mirai} (present only if using dispatcher) - a named integer
vector comprising: \strong{awaiting} - number of tasks queued for execution at
dispatcher, \strong{executing} - number of tasks sent to a daemon for
execution, and \strong{completed} - number of tasks for which the result has
been received (either completed or cancelled).
}
}
\description{
Retrieve status information for the specified compute profile, comprising
current connections and daemons status.
}
\seealso{
\code{\link[=info]{info()}} for more succinct information statistics.
}
\keyword{internal}
