% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{status}
\alias{status}
\title{Status Information}
\usage{
status(.compute = "default")
}
\arguments{
\item{.compute}{[default 'default'] character compute profile (each compute
profile has its own set of daemons for connecting to different resources).}
}
\value{
A named list comprising:
    \itemize{
    \item{\strong{connections}} {- number of active connections. Always 1L
    when using dispatcher as there is a single connection to the dispatcher,
    which in turn connects to the daemons.}
    \item{\strong{daemons}} {- if using dispatcher: a status matrix (see
    Status Matrix section below), or else an integer 'errorValue' if
    communication with the dispatcher was unsuccessful. If not using
    dispatcher: the host URL.}
    }
}
\description{
Retrieve status information for the specified compute profile, comprising
    current connections and daemons status.
}
\section{Status Matrix}{


    When using dispatcher, \code{$daemons} comprises a matrix with the
    following columns:
    \itemize{
    \item{\strong{online}} {- shows as 1 when there is an active connection,
    or else 0 if a daemon has yet to connect or has disconnected.}
    \item{\strong{instance}} {- increments by 1 every time there is a new
    connection at a URL. This counter is designed to track new daemon
    instances connecting after previous ones have ended (due to time-outs
    etc.). 'instance' resets to zero if the URL is regenerated by
    \code{\link{saisei}}.}
    \item{\strong{assigned}} {- shows the cumulative number of tasks assigned
    to the daemon.}
    \item{\strong{complete}} {- shows the cumulative number of tasks
    completed by the daemon.}
    }
    The URLs are stored as row names to the matrix.
}

\examples{
if (interactive()) {
# Only run examples in interactive R sessions

status()
daemons(n = 2L, url = "wss://[::1]:0")
status()
daemons(0)

}

}
