% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{stop_mirai}
\alias{stop_mirai}
\title{mirai (Stop)}
\usage{
stop_mirai(x)
}
\arguments{
\item{x}{a \sQuote{mirai} object, or list of \sQuote{mirai} objects.}
}
\value{
Logical TRUE if the cancellation request was successful, or FALSE if
  already cancelled or completed (or not using \sQuote{next} dispatcher). A
  return value of TRUE does not guarantee actual cancellation as it is not
  always possible to interrupt an ongoing evaluation.
}
\description{
Stops a \sQuote{mirai} if still in progress, causing it to resolve
immediately to an \sQuote{errorValue} 20 (Operation canceled).
}
\details{
Forces the \sQuote{mirai} to resolve immediately. Has no effect if the
\sQuote{mirai} has already resolved.

This function returns TRUE if the cancellation request was successful. In
such a case if the mirai is queued for execution, it is discarded, and if
in the process of execution on a daemon, an interrupt signal is triggered.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

m <- mirai(Sys.sleep(n), n = 5)
stop_mirai(m)
m$data

}

}
