% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{stop_mirai}
\alias{stop_mirai}
\title{mirai (Stop)}
\usage{
stop_mirai(aio)
}
\arguments{
\item{aio}{a 'mirai' object.}
}
\value{
The passed mirai (invisibly).
}
\description{
Stops a mirai if still in progress, causing it to resolve immediately to an
    'errorValue' 20 (Operation canceled).
}
\details{
Forces the mirai to resolve immediately. Has no effect if the mirai
    has already resolved.

    If cancellation was successful, the value at \code{$data} will be an
    'errorValue' 20 (Operation canceled). Note that in such a case, the mirai
    is aborted and the value not retrieved - but any ongoing evaluation in
    the daemon process will continue and is not interrupted.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

s <- mirai(Sys.sleep(n), n = 5)
stop_mirai(s)$data

}

}
