% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{stop_mirai}
\alias{stop_mirai}
\title{mirai (Stop)}
\usage{
stop_mirai(aio)
}
\arguments{
\item{aio}{a 'mirai' object.}
}
\value{
Invisible NULL.
}
\description{
Stops a mirai if still in progress.
}
\details{
Forces the mirai to resolve immediately. If the asynchronous
    evaluation was not yet complete, attempting to access its value at
    \code{$data} will return a 'miraiInterrupt' (see
    \code{\link{is_mirai_interrupt}}).

    Note that in this case, it is only guaranteed that the mirai value is not
    returned - any ongoing evaluation in the daemon process will continue and
    is not interrupted.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

s <- mirai(Sys.sleep(n), n = 5)
stop_mirai(s)

}

}
