% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.R
\name{stop_mirai}
\alias{stop_mirai}
\title{mirai (Stop)}
\usage{
stop_mirai(x)
}
\arguments{
\item{x}{a \sQuote{mirai} object, or list of \sQuote{mirai} objects.}
}
\value{
A logical value: \code{TRUE} if the cancellation request was
  definitely successful (task was awaiting execution and has been discarded),
  \code{NA} if a cancellation request was successfully sent (task was in
  execution, and will be interrupted), and \code{FALSE} if already completed
  or previously cancelled. Will always return \code{FALSE} if not using
  dispatcher.
}
\description{
Stops a \sQuote{mirai} if still in progress, causing it to resolve
immediately to an \sQuote{errorValue} 20 (Operation canceled).
}
\details{
In the case that the \sQuote{mirai} is in execution and an interrupt was
sent, the task or a portion of it may have completed before the interrupt is
received, and even then it is not always possible to immediately interrupt
evaluation (especially if it is compiled code).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

m <- mirai(Sys.sleep(n), n = 5)
stop_mirai(m)
m$data

}

}
