% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemons.R
\name{with_daemons}
\alias{with_daemons}
\alias{local_daemons}
\title{With Daemons}
\usage{
with_daemons(.compute, expr)

local_daemons(.compute, frame = parent.frame())
}
\arguments{
\item{.compute}{[default NULL] character value for the compute profile
to use (each has its own independent set of daemons), or NULL to use the
'default' profile.}

\item{expr}{an expression to evaluate.}

\item{frame}{[default parent.frame()] the frame (environment) to which the
daemons compute profile is scoped.}
}
\value{
The return value of \code{expr}.
}
\description{
Evaluate an expression using a specific compute profile.
}
\details{
Will error if the specified compute profile is not yet set up.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
daemons(1, dispatcher = FALSE, .compute = "gpu")
status()

with_daemons("gpu", {
  m <- mirai("running on gpu")
  status()
})

m[]
status()
daemons(0, .compute = "gpu")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
daemons(1, dispatcher = FALSE, .compute = "gpu")

gpu_func <- function(.compute = "gpu") {
  local_daemons(.compute)
  mirai("running on gpu")
}
m <- gpu_func()
m[]
daemons(0, .compute = "gpu")
\dontshow{\}) # examplesIf}
}
